move server logic to server.utils, create server.views module for http views and extract them from server.api and server.models
