+++ catanatron_server/catanatron_server/api.py
@@ -1,98 +1,37 @@
-import json
+from flask import Blueprint, request
 
-from flask import Response, Blueprint, jsonify, abort, request
-
-from catanatron_server.models import upsert_game_state, get_game_state
-from catanatron.json import GameEncoder, action_from_json
-from catanatron.models.player import Color, RandomPlayer
-from catanatron.game import Game
-from catanatron_experimental.machine_learning.players.value import ValueFunctionPlayer
-from catanatron_experimental.machine_learning.players.minimax import AlphaBetaPlayer
+from catanatron_server.views import (
+    get_game_view,
+    post_game_view,
+    stress_test_view,
+    post_action_view,
+)
 
 
 bp = Blueprint("api", __name__, url_prefix="/api")
 
 
-def player_factory(player_key):
-    if player_key[0] == "CATANATRON":
-        return AlphaBetaPlayer(player_key[1], 2, True)
-    elif player_key[0] == "RANDOM":
-        return RandomPlayer(player_key[1])
-    elif player_key[0] == "HUMAN":
-        return ValueFunctionPlayer(player_key[1], is_bot=False)
-    else:
-        raise ValueError("Invalid player key")
-
-
 @bp.route("/games", methods=("POST",))
 def post_game_endpoint():
     player_keys = request.json["players"]
-    players = list(map(player_factory, zip(player_keys, Color)))
 
-    game = Game(players=players)
-    upsert_game_state(game)
-    return jsonify({"game_id": game.id})
+    return post_game_view(player_keys)
 
 
 @bp.route("/games/<string:game_id>/states/<string:state_index>", methods=("GET",))
 def get_game_endpoint(game_id, state_index):
-    state_index = None if state_index == "latest" else int(state_index)
-    game = get_game_state(game_id, state_index)
-    if game is None:
-        abort(404, description="Resource not found")
-
-    return Response(
-        response=json.dumps(game, cls=GameEncoder),
-        status=200,
-        mimetype="application/json",
-    )
+    return get_game_view(game_id, state_index)
 
 
 @bp.route("/games/<string:game_id>/actions", methods=["POST"])
 def post_action_endpoint(game_id):
-    game = get_game_state(game_id)
-    if game is None:
-        abort(404, description="Resource not found")
 
-    if game.winning_color() is not None:
-        return Response(
-            response=json.dumps(game, cls=GameEncoder),
-            status=200,
-            mimetype="application/json",
-        )
-
-    # TODO: remove `or body_is_empty` when fully implement actions in FE
-    body_is_empty = (not request.data) or request.json is None
-    if game.state.current_player().is_bot or body_is_empty:
-        game.play_tick()
-        upsert_game_state(game)
-    else:
-        action = action_from_json(request.json)
-        game.execute(action)
-        upsert_game_state(game)
-
-    return Response(
-        response=json.dumps(game, cls=GameEncoder),
-        status=200,
-        mimetype="application/json",
-    )
+    return post_action_view(game_id, request)
 
 
 @bp.route("/stress-test", methods=["GET"])
 def stress_test_endpoint():
-    players = [
-        AlphaBetaPlayer(Color.RED, 2, True),
-        AlphaBetaPlayer(Color.BLUE, 2, True),
-        AlphaBetaPlayer(Color.ORANGE, 2, True),
-        AlphaBetaPlayer(Color.WHITE, 2, True),
-    ]
-    game = Game(players=players)
-    game.play_tick()
-    return Response(
-        response=json.dumps(game, cls=GameEncoder),
-        status=200,
-        mimetype="application/json",
-    )
+    return stress_test_view()
 
 
 # ===== Debugging Routes
